Check the minimum free stack left for user code (#50)
Check the minimum free stack left for user code: modify uPortTaskStackMinFree() so that if passed NULL it will return the minimum free stack of the current task.  Add a #define to all u_cfg_test_platform_specific.h files defining U_CFG_TEST_OS_MAIN_TASK_MIN_FREE_STACK_BYTES (currently 1024 * 5 in all cases) and then at the end of every set of automated tests print out the minimum free task space of the current task and check that it is at least U_CFG_TEST_OS_MAIN_TASK_MIN_FREE_STACK_BYTES.  This way we guarantee the space left for the user.

NRF53 needs to be given more main task stack to pass this test, now 2048 from the original 1024.

As part of this, update ESP32 automation so that the ESP-IDF build it actually listens when the main task stack size is changed and set the STM3F4 main stack size correctly (it was being specified in bytes when the native STM32F4 RTOS API takes it in words, hence it had four times as much as it needed).