Add an "intercept" API to the AT client. (#53)
This API allows a function to be hooked in after the stream (e.g. UART) receives data from a module but before that data is processed by the AT client and, in the other direction, after the AT client has prepared data to send but before it gives that data to the stream interface for onward transmission to the module.  The intercept functions get to buffer/delay/frame the flow as they wish.  This is required for implementation of the u-blox Security as a Service chip-to-chip encryption feature.