Fix for inactivity timeout seen in the UDP async sockets test (sockUdpEchoNonPingPong()) when running on cellular. (#62)
What was happening was that in the socket receive functions of u_cell_sock.c we call the data callback if new data has arrived during a read (as well as if a URC indicates that it has arrived asynchronously).  However the sockets interface is locked at that point so if, like in my tests, the data callback calls into the receive function _again_ the system will seize since the mutex is already taken.  It is pointless to call the data callback when a read is already being done so simply removing that call fixes the problem.

Also made a few other comment clarifications while I'm at it.