Automation changes/fixes: (#66)
Add two STM32F4 platforms to the automated tests and remove one.  The one that is removed is the "x" that would be 14: it wasn't actually required, the things it tests can be distributed between the next instance up and the C030-U201 instance.  Note that this results in a renumbering: the "x" that would have been instance 15 is reinstated to what it was (C030-R412M) and becomes instance 14 and those above move down.  The new one on the end becomes instance 21, a C030 board with a SARA-R5 fitted to it.

With the introduction of these new boards/modules a few weeks was spent running and re-running the tests to catch and fix stability problems: these changes are scattered throughout the (mostly test) .c files in this commit.

At the same time, the STM32F4 automation has had an overhaul.  Switched back to Open OCD because the ST command line tools are _really_ sensitive to losses on the SWV interface, basically non-functional on the SARA-R5 board which, since it has suffered solder reflow I think, is a bit noisy in this respect.  Open OCD is much better, provided the data rate on the SWV line is brought down from 2 MHz to 100 kHz.  2 MHz was probably on the high side anyway, 125 kHz is really too low however it has to be that way to use the C030-SARA-R5 board which is quite a useful test platform.

Bringing the logging rate down requires a little more slack in one of the port tests, hence the change there.

Also spotted that the tail-end of the debug stuff out of u_main.c was probably being run with the port initialised, since the tests do a uPortInit()/uPortDeinit() at either end.  This now fixed.

Oh, and also moved the ST tools to 1.4.0 (from 1.3.0) while I was fiddling around.

Finally, in the Python scripts change .kill() to .terminate() in the hope of reducing the likelihood of nrfjprog hanging and add the facility to reset a USB port, which we do when nrfjprog hangs.  These seems sensible things to do but they haven't made the problem go away, unfortunately.