Automation change only: *finally* resolve the issue with J-Link download failures on the NRF52 boards, plus a few tidy-ups along the way. (#76)
Details: when using J-Link for RTT logging, driven from a Python script, there is only ONE way to reliably shut J-Link down again and that is to keep a stdin pipe open to it and send it the string "exit\n" before terminating it. This is the official advice from SEGGER and experiment shows it to be true.  If you send J-Link SIGTERM/CTRL-C/CTRL-BREAK signals, or if you just terminate it, some percentage of the time (anything from 10% to 40% of the time) J-Link will hang the next time you try to start it on that board and only power-cycling the board will fix the problem; we tried everything, including programmatically resetting the PC's USB port, and nothing helped; the Segger engineer thought that the J-Link DLL had probably sent a partial command to the debug chip and the debug chip was stuck waiting for more. When this happens, if you take a J-Link log (command-line parameter -Log <Path>) the log will end with "Out of sync, resynchronizing..." and then nothing more.