NRF52 automation change: lock the NRF52/NRF53 platform during download and when initiating RTT logging. (#82)
When multiple parallel downloads/RTT-logging sessions are running at once and J-Link is started for RTT logging J-Link sometimes seems to "attach itself" to nrfjprog if it happens to be running (against a different debugger): J-Link prints "Process: nrfjprog.exe" rather than "Process: jlink.exe" when it starts up to do RTT logging and then never emits any RTT logging whatsoever.  NRF52 does obtain its platform lock so this _seems_ to occur when nrfjprog is in use by west (the Zephyr build tool, not the orientation) for download to NRF53.

Hence with this change NRF53 also waits for the platform lock and there's a hack in u_pull_request.py to make sure that NRF52 and NRF53 share a platform lock.  Could do this by having them both wait on the system lock instead but then they'd have to wait for the looooong time that ESP32 takes to sort itself out ('cos it can potentially do some "installations" it needs to be sure that it's got sole access to the system at the time).  On balance the hack, which is clearly marked, is better than wasting time.