Automation change only: re-enable instance 14. (#87)
There's something odd going on here: the SARA-R412M module on the C030 board appears to stop responding to the STM32F4 processor after about a week of running on the test system.  However, some time later, with no apparent change to anything, it returns to life.  So this is expected to fail in a week or two, then I will disable it again and try to find out what's wrong.