Add u-blox security (#88)
WARNING WARNING WARNING: this change modifies the order of the first few elements in the uNetworkConfigurationCell_t structure because it was filled in wrongly in a few places, likely because it was simply not very intuitive.  It is now more logically laid out and people are more likely to use it correctly.

But the main purpose of this commit is to add the u-blox security-as-a-service features: security sealing, end-to-end encryption (SARA-R412M/SARA-R5) and chip-to-chip security (SARA-R5).