Automation change only: re-add instance 14 (C030-R412M). (#97)
Turns out that the SARA-R412M module can get into a state where it does not respond to AT commands and stays that way until power is TOTALLY removed.  On the test system all C030 boards have batteries plugged into them as well as USB power (in order to meet the peak power requirements of a cellular modem) and there is normally no way to cut this battery power.  I have now wired a 5V relay in series with the battery leads of the battery plugged into instance 14, powered from the 5V output on that same C030 board, so that whenever the board is powered down (which is done before the test suite us run using a USB cutter) the battery will be disconnected also.