UART test pin configuration change but, mostly, NRF53 UART port change (#93)
UART test pin configuration change but, mostly, NRF53 change: on NRF53 Zephyr is used as the RTOS and with Zephyr the pin assignments for any peripheral managed by Zephyr (e.g. UART) is done at compile-time, not at run-time.  To avoid confusion (since this is unlike any other RTOS) this change adds a note explaining this to the uPortUartOpen() API call and, since only Zephyr is supported on NRF53, modifies u_port_uart.c for NRF53 to return an error if the "should be ignored" run-time pin assignments are anything other than -1 (the pins in the cfg header file for NRF53 were already set to -1 anyway).  The UART test in u_port_test.c is then modified so that it only runs the UART test with flow control off if either (a) the pins are compile-time configurable and the flow control pins are not connected or (b) the pins are run-time configurable.  Likewise, it only runs the UART test with flow control on if either (a) the pins are compile-time configurable and the flow control pins are *connected* or (b) the pins are run-time configurable.  This test ran perfectly fine before but the prints would show confusing information, for instance saying that it would ignore flow control when in fact it couldn't.

While doing this the _GET macros for test UART pin numbers are extended to be available for the TXD and RXD pins since that make things nice and symmetrical (and this facility is mentioned in a comment in the implementation of uPortUartOpen()).

A note is also added to the README.md for the Zephyr SDK as to where the pin settings can be found/modified at compile time.