Platform directory restructure, add NRF52 to Zephyr, Zephyr RTT logging (#120)
IMPORTANT: this is a major change to the port/platform directory structure.  We realised when adding support for NRF52, which is already supported under nRF5, to Zephyr that the thing upon which ubxlib's port implementation stands is the *SDK* of the chipset and *not* the chipset.  Hence there is a lot less duplication of configuration files etc. if the directory structure is organised the same way.  So the thing under platform is now the SDK (nRF5, STM32Cube, ESP-IDF or Zephyr) and the MCUs can be found deeper down.

Notes:
- this repo now uses sub-modules (to bring in some custom Zephyr board files) so make sure you do something like:
  `git submodule update --init --recursive`
  ...after cloning or updating it,
- all tracing for Zephyr is now sent over RTT just like the nRF5 case, freeing up the UART ports for testing; the README.md in zephyr/runner contains instructions for how to switch it back if you prefer trace over UART,
- the Zephyr platform directory structure differs from all the others because of the amount of board-specific configuration that Zephyr requires: hopefully these differences will be natural to Zephyr users,
- for anyone using the test automation scripts, the name of the settings file has been temporarily changed in order to allow existing settings to be changed without affecting other branches; the next PR will revert the name change.

Co-authored-by: Andreas Karlqvist <andreas.karlqvist@u-blox.com>