Test reliability change only
Follow-up to previous test reliability commit where I modified the check against 0 for the cellular data count to be a check for `<= 52` instead since a socket closure might occur asynchronously, after the counter reset.  I only changed the Tx counter part of the test I needed to also change the condition that follows 'cos the Rx count is also different in the same way when this condition occurs. (#126)