Test change only
Fix diagnostic print in portUartRequiresSpecificWiring() test.  Whether flow control is on or off in the test was being printed-out based on the values of the flow control pin #defines but that doesn't work for platforms where the pins cannot be set at run-time, need to do it based on the value of the flowControlOn parameter to the runUartTest() function, which is set based on discovering what the flow pins are according to platform-specific configuration mechanisms (e.g. in Zephyr). (#128)