Test change only: sockUdpEchoNonPingPong() reliability improvement (#131)
Test reliability improvement: in the sockets test sockUdpEchoNonPingPong() only print a warning (rather than performing a test assert) if the data callback wasn't called during the test. This is because, in the cellular sockets implementation, the data callback isn't called if the uSockReceiveFrom() or uSockRead() call is active when the data arrives (to avoid recursion) and this can, statistically, happen in this test since it calls uSockReceiveFrom() blindly without waiting for the data callback to be called.  There are other tests where the behaviour of the data callback is tested (the asynchronous ones).