nRF5SDK/GCC change only:  support GCC 10.2.1. (#133)
nRF5SDK/GCC change only: the latest available version of GCC (which the nRF5 SDK uses in its GCC flavour) is now 10 2020-q4-major and, in it, the version of newlib has changed from 3.0.0 to 3.0.3. heap_useNewlib.c for this platform has been tested with newlib versions 3.0.0 and 3.0.1 and so prints a warning since it does not recognise the newlib version string. Our nRF5DK/GCC make file is set up to treat warnings as errors and hence when this happens compilation fails.

With this commit heap_useNewlib.c is modified to add 3.0.3 as a supported newlib version for nRF5SDK/GCC and the default GCC version expected by nRF5SDK/GCC during automated testing is set to 10 2020-q4-major (paths in u_settings.py changed appropriately and so will take effect if your settings.json file is ever deleted).

Note: the STM32Cube version of heap_useNewlib.c is left unchanged since the version of GCC inside the STM32F4 board support package (STM32Cube_FW_F4) is not changed.