Cellular: fix uCellInfoGetRsrqDb() (#135)
Cellular fix for RSRQ: RSRQ always used to be a negative number, hence zero could be used for the error case.  However the latest version of TS 3GPP 36.133, Table 9.1.7-1, introduces an extended RSRQ range which permits small positive values (up to 2.5 dB).  SARA-R5 supports these values, hence the error value is now moved to be 0x7FFFFFFF.

As an aside, the coded value we read off the AT interface is now, for the first time, a negative number.  uAtClientReadInt() says it doesn't support negative integers but actually that's not true, the issue is that negative return values are also used to indicate an error.  The comment above uAtClientReadInt() is updated to explain how to work with that and read negative numbers correctly.