Test automation changes only (#127)
Automation change only: in an earlier change to the way test automation works some locks were added.  When an individual test instance is run from the command-line with u_run.py the locks are set to None ('cos they are irrelevant) but I wasn't checking for this and so Python got upset about trying to iterate over a NoneType; this fixes that.

Also, STM32F4 instance 16 has been failing to download occasionally and, looking at the logs, this just seems to be due to system loading at the time, it takes 30 seconds just to begin the download process, hence here we increase the download timeout from 60 seconds to 120 seconds.