Add example: C030 module FW update. (#139)
Add an example, which is actually more of a utility (hence the utilities sub-directory) that puts a C030 board into a state where a SARA-R4xx or SARA-R5 cellular module on that board can have its FW updated.  The main change here is actually in u_main.c for STM32F4 where the PWR_ON pin to the cellular module is set to be floating before the module is taken out of reset: this puts the module in a state where it can be firmware-updated.