Network API: better state definition and change of behaviour for cellular (#143)
Network API: better state definition and change of behaviour for cellular.

The network API now includes a more detailed description of each network state and guidance on how to use the API calls.  The single functional change is to the cellular implementation: uNetworkAdd() now powers-up the cellular module whereas previously that was only done at the start of uNetworkUp().  This means that uNetworkAdd() can be called and the cellular module will be available for configuration (e.g. in a future PR the security credentials stored/accessed) before uNetworkUp() makes the connection.