Security change (mainly C2C) plus additional API function in AT Client. (#147)
Security change (mainly C2C) plus additional API function in AT Client.

Fix three issues in the C2C implementation, switch the testing back on again and add an example of how to use C2C.  Note that the example only runs if U_CFG_TEST_SECURITY_C2C_TE_SECRET is defined as, otherwise, it could irreversibly C2C-pair a customer's module by accident.

In addition, the call to uSecurityGetRootOfTrustUid() is modified to permit a NULL buffer to be passed in: this way it can be used as a simple check that a module's u-blox security features are up and running.

And, finally, an API is added to the AT client to switch stop-tag detection back on.  Reason for this is that, when stop-tag detection is switched off (as it must be to receive binary data), the uAtClientResponseStop() function doesn't wait for OK or ERROR, it just flushes the buffer and returns; this is fine normally because the OK or ERROR will likely be in there but it doesn't work reliably in the C2C case: C2C is lumpy and the OK or ERROR might not quite have arrived yet so it is best to make sure that those AT sequences which call uAtClientIgnoreTopTag() have a way to switch stop tag detection back on again before calling  uAtClientResponseStop(), ensuring we wait for OK or ERROR before issuing the next AT command.

Oh, and the test automation script u_monitor.py is updated to throw away rubbish received on the Telnet interface rather than barf on it.