New platform: static_size. (#151)
New platform: static_size.

We were asked a few days ago how big ubxlib is so, to facilitate answering that question in future and, possibly more importantly, to also have a build of ubxlib without the test, example and platform code (i.e. to check that there is no header creep), this commit adds a new platform which simply compiles/links the core code standalone under ARM GCC (with optimisation for size) and prints out the static flash/RAM sizes.  This is a new instance in DATABASE.md which will always be run.