Cellular change only: fix intermittent TLS connection failure. (#168)
Cellular change only: a TLS-based socket connection attempt can fail on SARA-R5 if it is invoked immediately after the socket is secured; this commit introduces a 250 ms delay to prevent that. Without this change three connection failures were seen in 30 test runs, with this change zero connection failures were seen in 30 test runs.