Example code change only: add connection retry. (#170)
The sockets examples sometimes fail to connect when run under the test campaign, I believe due to collisions of connection attempts at the test echo server when establishing a secure socket.  It is sensible, in any case, to have a retry with a delay as connections can fail in the real world, so add a retry to the sockets examples.