Cellular, sockets, only change: add AT+USOERR for debug. (#172)
When a sockets operation fails on the AT interface the module can be queried for further error information with AT+USOERR.  With this commit, whenever an AT socket operation fails, this command is emitted for debug purposes (i.e. it is visible in the AT log if that is being printed but nowhere else).