Zephyr change only: switch to using minimal C lib instead of newlib. (#169)
Zephyr change only: switch to using minimal C lib instead of newlib.

It turns out that the Zephyr integration of the non-malloc() functions in newlib, i.e. snprintf(), printf(), scanf(), strtok(), etc., is not thread-safe.  This commit switches over to using the default Zephyr minimal C library instead.  This entails the addition of a number of functions in the port/platform/clib directory which the Zephyr minimal C library doesn't support: isblank(), mktime() and setjmp()/longjump().

Note that, unfortunately, the Zephyr minimal C library doesn't support any form of heap usage checking so we have to rely on other platforms for that.