Revert "Example code change only: add connection retry. (#170)" (#178)
While the change kind of made sense, in fact if there is to be a retry loop it is better across the whole socket create/connect/[close] process since failing to connect a socket can result in it being closed, hence it would need to be speculatively closed and re-created.  That is to messy for a simple example so go back to the original version.