Cellular change only: GPIO pull direction fix.
The way GPIO initialisation for the cellular pins was performed was messy and, in particular, resulted in a different pin configuration for the VInt pull up/down state depending on whether a pinEnablePower happened to have been defined. With this change the pin configuration is reset to default before each pin-setting and the VInt case is fixed to be "no pull".