Test automation change only: allow ESP-IDF system lock to be None. (#184)
Because the ESP-IDF Python script installs stuff it needs a system-lock to prevent different parallel instances from falling over each other. However, when run from u_run.py (as opposed to u_run_branch.py) there can be only one instance and so the system lock is None. This used to cause Python to barf; with this commit it no longer does, so a u_run.py of an ESP-IDF instance will now work.