Cellular security change: SARA-R5 certificate checking. (#181)
The default certificate checking level for TLS security in U_SECURITY_TLS_SETTINGS_DEFAULT is none. However, this is not the default setting within SARA-R5, which defaults to a root CA check, hence the settings are out of whack in this case. With this commit the default in U_SECURITY_TLS_SETTINGS_DEFAULT (i.e. none) is obeyed in all cases (SARA-R5 is configured that way when security is applied to it) and root CA checking is instead specifically switched on in the example code.