AT client and security test change related to C2C security. (#175)
There seems to be a circumstance, somewhere, one that occurs very rarely and has been seen on NRF52/NRF53 MCUs only, in which, under chip-to-chip encryption, the buffer lengths in the AT parser go wrong; specifically the length is larger than the length buffered, leading to a negative (i.e. very large positive) size being calculated.  This commit adds defences in the places where a negative length could be calculated, prints a warning when it happens and re-enables the C2C tests.

Also, detailed debug (under a conditional compilation switch) is added to the AT client and a test added to the security tests to try to (a) flush the problem out and (b) help debug it if it can be made to occur reliably.