Test automation change only: JLink shut-down time (#191)
JLink is quite peculiar in that it will ONLY exit tidily if you send it "exit\n" over stdin.  No other mechanism works (I have this direct from Segger).  After sending it this exit string we wait 5 seconds before assuming that it has ended and releasing the JLink lock so that the next thing can use it.  However it appears that under high machine loading this is not long enough for all of the logging to get out and into the test system; a new instance of JLink will just terminate the previous one and chop off the log prematurely.  Increasing the wait time to 10 seconds makes that problem go away.

This change also makes the time for shut-down a setting in the settings file so that all the scripts that use JLink can pick up a safe value.