Cellular optional change only: AT+CMEE=1
By default extended errors, AT+CMEE, is set to 2 to get the textual description of the error from the module, which is most beneficial when developing with the AT interface prints switched on.  However, this then necessarily omits the numeric error and so uAtClientDeviceErrorGet() can only return zero for deviceError.  This commit introduces a compilation flag, U_CFG_CELL_ENABLE_NUMERIC_ERROR: if this is defined AT+CMEE is set to 1 to obtain numeric errors only.