Cellular C2C test change only: more tolerant to speed variations. (#195)
The test cellSecC2cAtClient() runs two tasks either side of back-to-backed UARTs to run actual real tests of the chip to chip encryption code through the AT client as well. However, since neither task can be aware of the arriving data size (the encrypted size is not actually known), the test is somewhat sensitive to how long things take to get around the loop, which is of course platform dependent. Upcoming PR #194 brings BLE into the picture on the Zephyr platform and that seems to have added load to the underlying system somewhere so that things, maybe just UART things, possibly task time, it is difficult to tell, take a little longer, or maybe add jitter, causing this test to fail. This commit re-writes the test to hopefully improve its tolerance to perturbations in the neutron flow.