Security credentials change: add 64-bit mktime(), mktime64(). (#188)
The uSecurityCredentialListFirst()/uSecurityCredentialListNext() functions return the expiry date of a certificate as a UTC timestamp. These timestamps are now approaching the 32-bit wrap-around point of 19 January 2038 and our platforms only have time_t as a 32 bit number. We have recently added our own version of mktime(), for the case where we're using the Zephyr minimal library that does not have mktime(); with this commit that mktime() is split into a 64-bit version, mktime64(), and the normal version mktime(). mktime64() is made available in a header file of its own and it is used in u_security_credential.c to give a guaranteed 64-bit time value.