Test automation change only. (#196)
Test instance 0.2 runs u_run_lint.py in a mode where it checks that any "_stub.c"/"_stub.cpp" files are good. Unfortunately it had a mistake in it, using .rstrip() when it should actually use .replace(). This is benign until a filepath with any of the characters '_', 's', 't', 'u', or 'b' at the end appear with a "_stub.c"/"_stub.cpp" version. Now fixed.