Cellular change only to SARA-U2 RAT setting. (#200)
SARA-U2 requires that AT+COPS must be 2 (deregistered) as well as AT+CFUN being 1 when changing the RAT. This didn't used to matter but it seems it does now, guessing due to a change in the local cellular radio environment where the test farm is located. With this change, for SARA-U2 only, AT+COPS is switched to 2 for the RAT change and then put back to what it was again afterwards.