Platform change only: increase guard timer in static_size
The tests which invoke the static_size script sometimes fail on the test system due to hitting the guard timer. This is because, though 30 seconds seems a generous guard timer for compiling a single file, the test machine is extremely heavily loaded and can take a long time just to find the memory/CPU to start a process. With this change the guard timer is increased to 120 seconds.