Test automation: Move process prio elevation to u_utils.py
Changing process priority works a little diffrent in Linux compared
to Windows. Linux just use a number where 0 is the default prio,
a negative value for higher prio and a positive value for lower prio.

For Linux it should be possible to set prio with:
  psutil.Process().nice(-10)
However we get "[Errno 13] Permission denied" even when run as root.
I think the reason is that the user I tested with is missing the
Linux capability CAP_SYS_NICE. We can add this to the Docker container
later on, but for the time beeing I disable process priority elevation
for Linux.