Security credential list fix/addition to AT API: fix credential listing, multi-line response with no prefix. (#217)
The AT+USECMNG=3 command, used to list security credentials, is a peculiar one in that the response is multi-line and the lines do not have a prefix on them.  This means that it is impossible for the AT parser to tell whether it has received an information response to its command or a URC.  This is not a problem in most circumstances because the parser can detect the CR/LF on the end of a line and returns, giving the application the chance to decide how to handle the stuff it has received (i.e. either it has set a URC or it is actively reading the data for an expected sequence).  However, in the multi-line case it is intended that the application should be able to call uAtClientResponseStart() repeatedly to process multiple lines and this doesn't work reliably if the AT parser is unable to tell whether it was in an information response or not: it works if the CR/LF happens to fall inside the first run around the application's "line loop" but not otherwise.

Hence this PR introduces a new function to the AT parser, uAtClientIsInformationResponse() which can be called by the application in the specific situation where it intends to call uAtClientResponseStart() multiple times for a response where the lines include no prefix.  The function should be called by the application after it has determined that the response it has received really _is_ an information response and before calling uAtClientResponseStart() for the next line.

The function is used in uSecurityCredentialListFirst() and instance 12 (ESP32/SARA-R5), which showed the problem, now passes this test reliably.