Test change only: fix AT client test race condition. (#222)
When testing the AT command timeout a variable is set in an asynchronous call-back task and then checked in the main task.  Depending on RTOS tick rates and timing it is possible for the check to be carried out before the asynchronous task has had chance to set the variable.  This is fixed by adding a delay of U_CFG_OS_YIELD_MS in the main task between the condition being triggered and it being checked.