AT client change: handing of responses with no prefix. (#220)
IMPORTANT: This is a change to the behaviour of the AT client.  It has been tested and passes all the ubxlib tests but if you are using the AT client in something of your own you may wish to check that this has no adverse effect.

In the case of an AT command in which the response has no unique prefix uAtClientResponseStart() would return if it had been given some characters of a line but had not yet received the first line terminator. This caused a problem under some circumstances (in particular the un-prefixed multi-line response to the AT+SECMNG=3 command from any module to the ESP32 platform where the OS tick is relatively slow, as evidenced by the occasional failures in securityCredentialTest on instance 12) in which sometimes only a part of the first response line would be received: uAtClientResponseStart() would be called and would return, the subsequent calls to the AT parser would find nothing (because there wasn't enough received yet) and, on the next call to uAtClientResponseStart(), the "rubbish" would be flushed out to the end of the line, losing the wanted characters.

With this change uAtClientResponseStart() always waits until either the prefix has been received or, if there is no prefix, then at least a line terminator has been received (or there has been a timeout). The note above the definition of U_AT_CLIENT_BUFFER_LENGTH_BYTES in u_at_client.h always said this was the case but previously it wasn't a hard requirement; now it is.

The commit from a few days ago, in which an API was added to permit the AT parser to be told that it is in information response mode, is no longer required; this commit is the correct fix for the problem and, as we've not made a push to the public ubxlib repo since it was committed, it is safe to remove it here.