Automation change only: fix settings collision. (#227)
When u_settings.py detects that an agent-specific setting is missing from settings_v2_agent_specific.json, or finds some other problem with the settings, it re-writes the settings files to either (a) highlight the missing setting or (b) create the necessary default settings.  This works fine except that u_settings.py is imported into several different Python scripts that can be running at the same time in different threads/processes and hence the re-writing of the settings files can collide with them being read.  This commit uses portalocker to add a [platform-independent] lock, preventing the collision from occurring.