Lint change only: workaround for intermittent automation issue. (#231)
In the automated test system Lint will sometimes fail because "make" is unable to delete a temporary file.  It is not clear why this is occurring, could maybe be the virus checker or something in Windows clinging-on to an executable file (which is one of the things being deleted).  Whatever, adding a 1 second delay after the executable file has finished running and before attempting to delete it is the only workaround we have found.