Add cellular power-on pin inversion option. (#224)
The PWR_ON pin to u-blox cellular modules is an input which is normally pulled high by the module and can be pulled low for a given period by the MCU to effect power on or power off.  However some boards (e.g. the Sparkfun Asset Tracker board, which includes a SARA-R5 module) have an inverter between the MCU pin and the module PWR_ON pin it is connected to.  For these cases the conditional compilation flag U_CELL_PWR_ON_PIN_INVERTED is introduced; #define it to invert the sense of the PWR_ON toggle.

The same is true of the VINT pin, an output from a u-blox cellular module which goes high to power level shifters and can be connected to a pin of the MCU as a means of detecting that the module is on, sometimes via an inverter.  For this the conditional compilation flag U_CELL_VINT_PIN_INVERTED is introduced; #define it to invert the sense of the VINT signal as seen by this code at the MCU.