Automation change only: don't re-write the settings files if they exist. (#232)
Currently if a settings file cannot be JSON decoded it is re-written with defaults. This is a right pain if you just happened to insert a JSON syntax error in the file and the system goes and overwrites all the stuff you put in there. With this change, if the file exists but cannot be read, it won't be over-written.