Test application change only: correct NRF5 SDK/RTT logging workaround. (#234)
In the previous commit a delay was added to the startup of the NRF5 SDK test application to permit RTT logging to connect to the MCU before the [recently enlarged] list of tests was printed, since it appeared that sending a large lump of RTT logging without anything connected to the target was causing the NRF5 SDK RTT logging system to become irretrievably blocked.

However, the complete test pass that fix achieved, it seems, was a fluke because the failure came back the moment the commit was merged into master. After much experimentation the only fix that works is to increase the NRF5 SDK logging buffer from 512 bytes to 1024 bytes. Hence this fix reverts the addition of the delay and increases the buffer size instead. For good measure, this fix has been through the complete suite of all tests five times and, in each case, the problem did not occur; fingers crossed...