Amendment to earlier Zephyr BREAKING CHANGE commit: correct BOARD_ROOT syntax.
Found when adding the cellular board [see next commit]: turns out that the separator for entries in the BOARD_ROOT needs to be ";" instead of " ".  Also correct Zephyr board list in u_settings.py (path needs to be to be to the directory _above_ board) and use forward slash in all of the Zephyr settings for consistency.