Add first Zephyr cellular board: Sparkfun Asset Tracker with NRF52840 MicroMod module.
The board definitions included are sufficient for testing in the ubxlib test farm.  They don't pretend to be a proper/complete board definition, though they may be a good "jumping off point" if you want to make one.  Note that the instances in the automation system here don't currently build/run this board, the board definition file was tested on the MKII automation system development branch.