BLE test change only: correct deinit order. (#244)
In the BLE test code uShortRangeEdmStreamClose() was being called after uAtClientDeinit(), however uShortRangeEdmStreamClose() calls into the AT client to remove the transmit intercept function, which results in the code accessing free()ed memory. This commit switches the order around. [The issue was spotted while preparing to move to a later version of the ESP-IDF tools where it happens to cause an abort() inside newlib lock() 'cos the mutex was inside the free()'d block].