ESP-IDF change: move to ESP-IDF master, BREAKING CHANGE FOR ubxlib TEST SYSTEM USERS. (#246)
IF YOU ARE USING THE ubxlib TEST SYSTEM it is best to delete and then re-populate the ESP-IDF directories manually before running with this code version, just to make sure it is all good, as jumping whole release versions can be a bit trying on the poor u_run_esp_idf.py script.  On Windows, and assuming the ESP32 tools are under c:\esp32, to do this delete c:\esp32\esp-idf-latest and then, from a freshly opened Administrator command prompt:

cd c:\esp32
git clone https://github.com/espressif/esp-idf esp-idf-latest
cd esp-idf-latest
git checkout origin/master
git submodule update --init --recursive

Test that it can install with:

set IDF_TOOLS_PATH=c:\esp32\esp-idf-tools-latest
install

Best not to use that command prompt for anything else afterwards as it is now full of ESP-IDF environment and will, for instance, point to their Python installation.

Detail:

Up to now the ubxlib test system has used two versions of ESP-IDF: for the Espressif boards release 4.1 and, for a u-blox NINA-W1 board, a fork of the ESP-IDF master branch from around two years ago with a change made in spi_flash_chip_generic.c to accommodate the flash chip used in the NINA-W1 module (which was being unwarrantedly marked as "unsupported" by the ESP-IDF code at that time and was never fixed in release 4.1).  However, we've recently changed the test system to not dump any build products into the code directory and that requires being able to override the environment variable SDKCONFIG in CMakeLists.txt and that variable doesn't exist in the ESP-IDF master branch that we happened to take for the NINA-W1 mod.

ESP-IDF has since been changed to be much less "rejecty" of flash chips so we can now move everything to ESP-IDF master, which was always the intention.