Automation change only: further work on controller-agent pattern part1a (#247)
Automation change only: further work on controller-agent pattern, part 1a.

Further to part 1, in this commit the controller/agent scripts are updated but again Jenkinsfile and DATABASE.md are unchanged and hence the current automated test system will continue to work as it did before. The following changes are made:

- introduce u_process*.py which allow test runs to be stopped in a controlled way as a result of a Jenkins abort, ensuring that processes finish in their own time (by experiment taking no more than around 60 seconds to do so worst case), rather than being "crow-barred" by Jenkins, and any completed log files etc. are still archived,
- no longer creates any files under the ubxlib directory,
- passes the list of changed files to u_controller_client.py via a file in order to reduce the command-line length,
the u_agent_service.py -d parameter will now cause it to maintain a windowed 10,000 line log file (i.e. this can be left on all the time) for debugging purposes,
- agent_start.bat gets a parameter /k which tells it to kill every running Python process before it starts; good for clean-up,
u_controller_client.py gets a parameter -o which will cause it to [also] write its output to a file and that will be archived along with everything else,
- all log files now have default .txt extensions as that allows Firefox to display them in the browser window rather than having to download them,
- sleep(0.01)'s added to u_monitor.py to ensure that abort requests are handled in a timely manner,
- the build stages of all instances, as well as the "run" stages, now react to abort requests.